requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.3
django-filter==23.2
django-heroku==0.3.1
gunicorn==21.2.0
packaging==23.1
psycopg2==2.9.6
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
